Added Makefile that makes the register and test fixtures Also moved fix features into root directory
