+++ fixture-features/duplicate-channel-names.js
