+++ fixture-features/fine-positions.js
@@ -1,4 +1,4 @@
-const FineChannel = require('../../lib/model/FineChannel.js');
+const FineChannel = require('../lib/model/FineChannel.js');
 
 module.exports = [
   {
