+++ fixture-features/multiple-categories.js
