+++ fixture-features/multiple-focuses.js
