+++ fixture-features/switching-channels.js
